+++ smartcontract/contracts/libraries/AccountsLibrary.sol
@@ -18,27 +18,39 @@ library AccountsLibrary {
         delete self._users[key].accessList;
     }
 
-    function setAdmin(Accounts storage self, address key) internal{
+    function setAdmin(Accounts storage self, address key) internal {
         self._users[key].isAdmin = true;
     }
 
-    function resetAdmin(Accounts storage self, address key) internal{
+    function resetAdmin(Accounts storage self, address key) internal {
         self._users[key].isAdmin = false;
     }
 
-    function checkAdmin(Accounts storage self, address key) internal view returns (bool){
+    function checkAdmin(Accounts storage self, address key)
+        internal
+        view
+        returns (bool)
+    {
         return self._users[key].isAdmin && self._users[key].isValid;
     }
 
-    function addHill(Accounts storage self, address hillOwner, address key) internal {
+    function addHill(
+        Accounts storage self,
+        address hillOwner,
+        address key
+    ) internal {
         return self._users[key].accessList.push(hillOwner);
     }
 
-    function removeHill(Accounts storage self, address hillOwner, address key) internal {
-        uint len = self._users[key].accessList.length;
-        uint idx = findIndexOfHill(self,hillOwner,key);
+    function removeHill(
+        Accounts storage self,
+        address hillOwner,
+        address key
+    ) internal {
+        uint256 len = self._users[key].accessList.length;
+        uint256 idx = findIndexOfHill(self, hillOwner, key);
 
-        if(idx >= len){
+        if (idx >= len) {
             return;
         }
 
@@ -46,9 +58,13 @@ library AccountsLibrary {
         self._users[key].accessList.pop();
     }
 
-    function findIndexOfHill(Accounts storage self, address hillOwner, address key) private view returns (uint) {
-        uint len = self._users[key].accessList.length;
-        for (uint i = 0; i < len; i++) {
+    function findIndexOfHill(
+        Accounts storage self,
+        address hillOwner,
+        address key
+    ) private view returns (uint256) {
+        uint256 len = self._users[key].accessList.length;
+        for (uint256 i = 0; i < len; i++) {
             if (self._users[key].accessList[i] == hillOwner) {
                 return i;
             }
@@ -57,11 +73,19 @@ library AccountsLibrary {
         return len;
     }
 
-    function getUser(Accounts storage self, address key) internal view returns(User memory){
+    function getUser(Accounts storage self, address key)
+        internal
+        view
+        returns (User memory)
+    {
         return self._users[key];
     }
 
-    function exists(Accounts storage self, address key) internal view returns (bool) {
+    function exists(Accounts storage self, address key)
+        internal
+        view
+        returns (bool)
+    {
         return self._users[key].isValid;
     }
 }
