+++ smartcontract/contracts/libraries/AntHillsLibrary.sol
@@ -17,25 +17,39 @@ library AntHillsLibrary {
         delete self._hills[hillOwner].accessList;
     }
 
-    function addUser(AntHills storage self, address key, address hillOwner) internal {
+    function addUser(
+        AntHills storage self,
+        address key,
+        address hillOwner
+    ) internal {
         self._hills[hillOwner].accessList.push(key);
     }
 
-    function removeUser(AntHills storage self, address key, address hillOwner) internal {
-        uint len = self._hills[hillOwner].accessList.length;
-        uint idx = findIndexOfUser(self,hillOwner,key);
+    function removeUser(
+        AntHills storage self,
+        address key,
+        address hillOwner
+    ) internal {
+        uint256 len = self._hills[hillOwner].accessList.length;
+        uint256 idx = findIndexOfUser(self, hillOwner, key);
 
-        if(idx >= len){
+        if (idx >= len) {
             return;
         }
 
-        self._hills[hillOwner].accessList[idx] = self._hills[hillOwner].accessList[len - 1];
+        self._hills[hillOwner].accessList[idx] = self
+            ._hills[hillOwner]
+            .accessList[len - 1];
         self._hills[hillOwner].accessList.pop();
     }
 
-    function findIndexOfUser(AntHills storage self, address key, address hillOwner) private view returns (uint) {
-        uint len = self._hills[hillOwner].accessList.length;
-        for (uint i = 0; i < len; i++) {
+    function findIndexOfUser(
+        AntHills storage self,
+        address key,
+        address hillOwner
+    ) private view returns (uint256) {
+        uint256 len = self._hills[hillOwner].accessList.length;
+        for (uint256 i = 0; i < len; i++) {
             if (self._hills[hillOwner].accessList[i] == key) {
                 return i;
             }
@@ -44,30 +58,55 @@ library AntHillsLibrary {
         return len;
     }
 
-    function storeFileMeta(AntHills storage self, string calldata filename, string calldata cid, uint filesize, address hillOwner) internal {
+    function storeFileMeta(
+        AntHills storage self,
+        string calldata filename,
+        string calldata cid,
+        uint256 filesize,
+        address hillOwner
+    ) internal {
         self._hills[hillOwner].files[filename].cid = cid;
         self._hills[hillOwner].files[filename].size = filesize;
         self._hills[hillOwner].files[filename].isValid = true;
     }
 
-    function loadFileMeta(AntHills storage self, string calldata filename, address hillOwner) internal view returns (File storage) {
+    function loadFileMeta(
+        AntHills storage self,
+        string calldata filename,
+        address hillOwner
+    ) internal view returns (File storage) {
         return self._hills[hillOwner].files[filename];
     }
 
-    function resetFileMeta(AntHills storage self, string calldata filename, address hillOwner) internal {
+    function resetFileMeta(
+        AntHills storage self,
+        string calldata filename,
+        address hillOwner
+    ) internal {
         delete self._hills[hillOwner].files[filename];
     }
 
-
-    function getHill(AntHills storage self, address hillOwner) internal view returns (AntHill storage) {
+    function getHill(AntHills storage self, address hillOwner)
+        internal
+        view
+        returns (AntHill storage)
+    {
         return self._hills[hillOwner];
     }
 
-    function exists(AntHills storage self, address hillOwner) internal view returns (bool) {
+    function exists(AntHills storage self, address hillOwner)
+        internal
+        view
+        returns (bool)
+    {
         return self._hills[hillOwner].isValid;
     }
 
-    function exists(AntHills storage self, string calldata filename, address hillOwner) internal view returns (bool){
+    function exists(
+        AntHills storage self,
+        string calldata filename,
+        address hillOwner
+    ) internal view returns (bool) {
         return self._hills[hillOwner].files[filename].isValid;
     }
 }
@@ -81,6 +120,6 @@ struct AntHill {
 
 struct File {
     string cid;
-    uint size;
+    uint256 size;
     bool isValid;
 }
